# [REVIEW] Optimize `cudf.concat` for `axis=0`